Added proper logic for comment count for given status

+++ purple-server/src/main/java/com/bbytes/purple/rest/dto/models/StatusDTO.java
@@ -35,6 +35,6 @@ public class StatusDTO implements Serializable {
 	
 	private String time;
 	
-	private int commentCount;
+	private long commentCount;
 
 }
